backend/requirements.txt
flask
pylint
pymysql
Flask-Script==2.0.5
coverage==4.3.4
sqlalchemy
blueprints
Flask-SQLAlchemy
mysqlclient
sqlacodegen
uwsgi
flask-mysql
bcrypt
pyjwt
flask-httpauth
twilio
imgurpython
configparser
